requirements.txt
# Core dependencies
openai>=1.0.0
nltk>=3.8.1
python-dotenv>=1.0.0
tenacity>=8.2.3

# Testing and QA
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
hypothesis>=6.82.6
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
bandit>=1.7.5

# Type stubs
types-requests>=2.31.0.2
types-setuptools>=68.1.0.0 